# Reduce tds buffer size to match packetsize